Move cache files
