+++ src/Disks/ObjectStorages/S3/parseConfig.h
@@ -6,7 +6,6 @@
 
 #include <aws/core/client/DefaultRetryStrategy.h>
 #include <IO/S3Common.h>
-#include <Disks/DiskCacheWrapper.h>
 #include <Storages/StorageS3Settings.h>
 #include <Disks/ObjectStorages/S3/ProxyConfiguration.h>
 #include <Disks/ObjectStorages/S3/ProxyListConfiguration.h>
@@ -14,7 +13,6 @@
 #include <Disks/DiskRestartProxy.h>
 #include <Disks/DiskLocal.h>
 #include <Disks/ObjectStorages/DiskObjectStorageCommon.h>
-#include <Common/FileCacheFactory.h>
 
 
 namespace DB
