+++ src/Interpreters/Cache/FileCacheFactory.cpp
